third_party/blink/web_tests/platform/generic/fast/xmlhttprequest/xmlhttprequest-default-attributes-expected.txt

This test verifies that XMLHttpRequest objects has correct default value for its attributes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS xhr.onabort is null
PASS xhr.onerror is null
PASS xhr.onload is null
PASS xhr.onloadstart is null
PASS xhr.onprogress is null
PASS xhr.onreadystatechange is null
PASS xhr.readyState is 0
PASS xhr.upload.onabort is null
PASS xhr.upload.onerror is null
PASS xhr.upload.onload is null
PASS xhr.upload.onloadstart is null
PASS xhr.upload.onprogress is null
PASS successfullyParsed is true

TEST COMPLETE
